Initialise StringInfoData once for each worker
This fixes memory leak for very long running workers.
The StringInfoData was allocated in each loop iteration, but was never
freed.
To fix this, allocate it once before the loop and use it inside.
The memory is still not freed, but there's no need as the worker is
about to exit anyway after the loop ends.